requirements.txt
 requests
 flask
 pillow
 google-cloud-storage
 torch
 transformers
#hashlib
 #pickle
 # -e .